requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.2.1
transformers==4.51.3
torch==2.7.0
torchvision==0.22.0
accelerate==1.6.0
sqlalchemy==2.0.40
aiosqlite==0.21.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
twilio==9.5.2
pyserial==3.5
paho-mqtt==2.1.0
jinja2==3.1.6
ultralytics==8.3.114
